Client removal is now working reliably. Also simplified intern_atom
There was an annoying bug while developing this that was harder to track
down than it should have been. (Attempting to cycle to the next client
once all clients on a Workspace had been removed resulting in an
indexing error into the workspace clients array). I need some tests for
specific pieces of code but anything relying on the X server is going to
be a complete pain to mock out...need to look at how other WMs handle
this!
I've also made the choice to just kill the process if we fail to intern
atoms for some reason (similar behaviour for most of the xcb calls)
rather than attempt to recover at the moment. This is annoying as it
results in the entire X session dying if something fails but it prevents
weird state from propagating with the internal penrose state not
matching up to the real X11 state. Probably want to be smarter about
this sort of thing in the future.