Removing mocking library as a test dependency
I was trying to find a way to mock out the connection to the X server
and then spy on what X event were sent. It's going to be a lot cleaner
and simpler to just ignore the XConn entirely and just ensure the
correctness of the methods on the Workspace struct I think.