Improving test coverage
Currently test fail due to a bug with workspaces losing the focused
client when they are mapped/unmapped. The existing implementation of
having Workspace structs own the Clients themselves introduces far to
much complexity in the end it seems. Need to go back to a central client
list and having Workspaces as containers that hold IDs and yield actions
to the WindowManager for execution.