Fixing bug with tracking client workspaces
The issue with removing a client but the layout being broken was down to
the fact that removing the client is an X action that is not tightly
coupled to removing the client from the workspace. This is correct
though: we may want to remove a client from a workspace but not kill the
program. We are now correctly tracking which workspace a client is on
when removing it entirely but it feels like there is now a fair amount
of duplication in that state which needs sorting out.