Client dragging now kind of works
With the caveat that dragging forward from index 0 results in the
flip-flopping between the first to clients on the workspace (and
possibly permuting the others? Not sure yet...) I have test cases at the
Ring and Workspace level that seem to show that this shouldn't be
happening so perhaps it's happening in the WindowManager itself? Not
sure how though...