Client dragging and cycling now warp the cursor to the focused window
This resolves the issue with client dragging breaking due to focus being
stolen by the cursor as you drag through the layout position containing
the cursor itself. I initially had this with warping whenever an
enter_notify was received but that resulted in some rather horrifying
bugs around the cursor spinning through all active clients...maybe only
if the client is being created?