issue-2 adding 'paper' layout and support for layout config
In order to get paper to work, we need the layout to be triggered each
time focus is gained by a new client, we also need to ensure that this
does not result in an infinite loop of constantly recomputing layouts
when wrapping from one side of the screen to the other. In theory this
should be doable but try as I might, there were a lot of issues with
getting that working smoothly without the layout function itself having
to handle it. As a compromise, enabling 'follow_focus' for a given
layout will enable this behaviour of triggering on focus change but at
the cost of preventing wrapping clients when cycling focus or dragging
through the stack. This prevents the edge case from arising at the cost
of a slightly more cumbersome experience if you want to move a client
from one end of the stack to the other.
Since I was already adding that in, we now support floating layouts
(though no mouse support for moving a client yet) and the ability to
disable gaps on chosen layouts if wanted.