issue-9: Ignoring SIGCHILD
In the initial attempt at resolving issue-6 (creation of zombie
processes due to not waiting on spawned children) I tried explicitly
waiting on the spawned procs coming from key bindings. This felt less
wide ranging than explicitly ignoring SIGCHILD and works for processes
that we spawn but still leaves the issue of reparented processes not
getting reaped. issue-9 provides some examples of how other WMs handle
this issue and it looks like explicitly ignoring SIGCHILD is the
standard approach and probably more in line with what users are
expecting.
This also then reverts the change made as part of issue-6 that modified
FireAndForget to return and Option<Child> as this is no longer required
and adds line noise for internal only bindings. There was also an issue
with this not behaving correctly with bindings that spawn multiple
processes.