WIP: Adding initial impls for status bar
This is not ready for real use yet but it is now possible to render
something that _looks_ like a status bar at least. Need to add more
sensible handling of Fonts (centralised rather than in each styling),
hooks to allow for update, partial update and async updates driven by
Widgets that are not driven by hooks (e.g. clock, battery etc)
This also brings in 'anyhow' which I was looking at during the
prototyping. I'm going to make more use of that in the main APIs going
forward but for now it is limited to the rendering logic only.