Fixing issue with gen_keybindings and run_internal
We were not explicitly casting the Boxed closure returned by
run_internal to FireAndForget, as we were with run_external. When a
run_external binding was provided first, it allowed the compiler to
correctly cast all subsequent closures to FireAndForget: when
run_internal was first, the compiler tried to cast everything to _that_
closure type. I've not added a test for this as the macro relies on
running xmodmap and an X server which is a no-go in CI (or at least I
_think_ it is). The minimal example config has now been updated though
and the builds and runs fine locally.