BUG-29 Overwritten status bar
This is actually a more fundamental issue that simply overwriting the
status bar: looks like I have been subscribing to the wrong events for
window mapping (MapNotify instead of MapRequest). This change means that
XcbConnection now subscribes to the correct events (mirroring dwm
behaviour of only subscribing to MapRequest) and that windows are
positioned via layout on first before their initial mapping.
Almost as a side effect, this fixes the original issue with the status
bar being overwritten which was happening by the window being occluded
by the newly mapped clients.