Adding client_to_screen methods for WindowManager
Also fixes a bug in moving clients between workspaces / screens when
source and destination are both active on a monitor. This was previously
only triggering layout for the source screen, requiring a flip/flop of
active workspace on the destination to show the newly moved client.

Additional unit tests also added for Selector based methods in
WindowManager.