Adding ability to modify where clients are inserted when added
As requested in issue #39, this adds the ability to set a client insert
point on the WindowManager for use in future operations where clients
are added to workspaces. Currently this logic is used when a client is
created for the first time and when a client is moved to another
workspace. Rather than using a Selector, this makes use of a new utility
enum on Rings again that specifies an index as first, last, current
focus or a specific index.