Issue-78 Dialog windows now float by default
I thought that this _had_ been implemented as part of the
window_should_float impl for XcbConnection but the window type was being
read as a string rather than a u32 atom and then compared against the
atom _names_ rather than _values_, meaning that the auto-float types
never matched. This was missed originally because the auto-float classes
are correctly checked as strings and this was working fine.