Initial fix for #85
Looks like the check for whether or not certain window types should be
managed or not was only present in some unused code that was intended
for detecting clients on restart / startup. This is now also done on
each map request to make sure we don't accidentally take ownership of
something that should be left alone.