WindowManager now takes a Box<dyn XConn> instead of &'a dyn XConn
This really should have been how this was handled from the start but
I've been learning more about generics in Rust as I go and overlooked
this in the original implementation. Now that breaking changes are being
made in advance of 0.2.0 this is finally being tidied up!

This has several benefits:
  - No more <'a> constraint on WindowManager
  - Helper functions that construct an impl of XConn and then build the
    resulting WindowManager are now possible because the XConn doesn't
    need to be created at the top level.

I tried making the WindowManager itself generic over <C: XConn> but that
ends up leading to a whole host of issues around implementing
Draw::Widget impls and Hook impls due the need for the type parameter to
be bubbled up through to them (as they take WindowManager as arguments
in most of their methods). I think having WindowManager take Box<XConn>
is the right way to go as WindowManager isn't _really_ generic over the
XConn implementation: it's behaviour (as far as other code interacting
with it is concerned) should be the same. This way, we don't provide
wiggle room for Hooks / Widgets trying to do funky things based on the
underlying XConn implementation.

Arguably we should? But the re-write required to put that in place is
massive at this stage...