Embedding Config in WindowManager and removing &'static where possible
This is a bit of a horrifying commit...It has grown out of wanting to
simplify the use of config in the WindowManager by storing the user
provided Config struct directly in WindowManager itself. This involved
breaking the hooks out of Config, and removing reference types...so I
did that elsewhere as well...and then I tidied up a lot of the hooks and
widgets to accept things that impl Into<String> so that things are a bit
more flexible.

Long story short, refactoring on Boxing Day leads to forgetting to
commit little and often!