Fix for bug introduced in #88
Turns out I was misunderstanding the XCB API when doing the rewrite
earlier. I thought I had a bug of always creating new windows on screen
index 0 (which _was_ working) and tried instead to create windows with
the screen index that was requested. This blows up because there is a
difference between 'screen' as far as X / XCB is concerned and 'output'
as far as randr is concerned. At the bare XCB level, there is one
screen and then randr splits that into several outputs that can be
handled independently.

If I'm honest, I'm still not sure that I fully grok how multi-monitor is
supposed to be handled in XCB/randr but this commit fixes the startup
crash we were seeing and statusbars are now working correctly when using
two monitors on startup.