Making WindowManager generic over XConn
This is a pretty big, breaking API change: instead of taking a
Box<dyn XConn>, WindowManager is now generic over the XConn it is passed
on construction. This has a knock on effect for _everything_ that takes
a WindowManager as an argument (hooks, bindings) and then anything that
consumes those etc. This makes a few things a bit more verbose (need to
work on tidying those up) but has three primary benefits:
  1) Code that interacts with the XConn in some way is now clearly
     identified in the type signature (no more hiding inside of the
     WindowManager inside of a box.
  2) No more dynamic dispatch via dyn XConn.
  3) The XConn is now serializable directly rather than needing to be
     injected again on deserialization.

As an additional factor: XConn impls need to be serializable for
WindowManager to be serializable now. This needs to be done for the XCB
based impls still.