Moving Widget back to using dyn DrawContext for its methods
This preserves the current (pre-"generic WindowManager") behaviour of
being able to embed widgets inside of one another to allow the
functionality to be re-used. Requiring the generic over DrawContext to
be explicit results in needing to use PhantomData in any Widget that is
going to be embedded so that the parent Widget and ensure that the same
DrawContext impl is used for both. There are alternatives to this but
they are all equally unergonomic.
This approach (the original one) results in the use of & dyn DrawContext
which is annoying but not the end of the world.