Make WindowManager generic over XConn (#100)
* Making WindowManager generic over XConn

This is a pretty big, breaking API change: instead of taking a
Box<dyn XConn>, WindowManager is now generic over the XConn it is passed
on construction. This has a knock on effect for _everything_ that takes
a WindowManager as an argument (hooks, bindings) and then anything that
consumes those etc. This makes a few things a bit more verbose (need to
work on tidying those up) but has three primary benefits:
  1) Code that interacts with the XConn in some way is now clearly
     identified in the type signature (no more hiding inside of the
     WindowManager inside of a box.
  2) No more dynamic dispatch via dyn XConn.
  3) The XConn is now serializable directly rather than needing to be
     injected again on deserialization.

As an additional factor: XConn impls need to be serializable for
WindowManager to be serializable now. This needs to be done for the XCB
based impls still.

* making XcbConnection / XConn serializable and hydratable

* adding feature flag for serde in tests

* running test suites in github actions with/without serde

* adding accessors for the underlying XConn impl to WindowManager

* splitting apart the Widget trait so that DrawContext and XConn can be concrete per impl

* adding methods for XCB backed WindowManager instances

* removing redundant inner module

* fixing broken doc links

* Moving Widget back to using dyn DrawContext for its methods

This preserves the current (pre-"generic WindowManager") behaviour of
being able to embed widgets inside of one another to allow the
functionality to be re-used. Requiring the generic over DrawContext to
be explicit results in needing to use PhantomData in any Widget that is
going to be embedded so that the parent Widget and ensure that the same
DrawContext impl is used for both. There are alternatives to this but
they are all equally unergonomic.
This approach (the original one) results in the use of & dyn DrawContext
which is annoying but not the end of the world.

* fixing docs