Initial work on allowing Results in all user code (#103)
* Initial work on allowing Results in all user code

The main change here is making Hooks and Key / Mouse Bindings fallible.
This has the knock on effect of meaning that public methods on the
WindowManager that are intended for use as Keybindings need to return a
Result in all cases (a little disgusting) and that all Hook method impls
must return a Result. This currently leads to a lot of Ok(()) all over
the place so there is probably more tidying up to do.

While this is all going on, the XConn trait (and Draw?) should probably
also be more honest about its failure cases. At the moment, failures are
silent or swallowed in odd ways in a lot of places. The nice thing is
that the impact on user config / main.rs level code is fairly minimal so
far. More so if users are writing their own widgets / hooks / trait
impls but that is to be expected.

* adding missing test import when using default features

* parsing keypresses now returns a result. Also provide non-blocking version

* adding non_blocking XConn functionality