ClientMessage: Use array instead of Vec
This commits changes the data member of ClientMessage from type Vec<u32>
to [u32; 5]. This new type captures the necessary invariants better and
also saves an indirection (although I bet this is an unmeasurable
improvement).

All existing callers of from_data_unchecked() already used an argument
of type &[u32; 5] which was automatically coerced to &[u32]. Thus, the
API change here has no consequences (and since it is pub(crate), there
are no consequences outside of penrose).

Signed-off-by: Uli Schlachter <psychon@znc.in>