issue #129 - improving handling of which windows are managed
Penrose still doesn't correctly manage all window types but this seems
to allow for sorting out windows that don't set a window type (such as
'st' or 'mpv') and also prevents transients and 'unviewable' windows
from being mapped on restart.

I suspect that handling around floating clients is still wonky but I
need to get some clean reproducible test cases to work with. Firefox is
behaving very strangely when being closed at the moment and I'm not sure
if it's due to how penrose is closing the window (only Firefox seems
affected) or if it is something wrong with my system...