Adding tests for the validate_user_bindings macro
This also makes sure that all test suites in the workspace are run in CI
and under 'make check-all' locally. There was an issue with one of the
doc-tests in penrose_menu which has been addressed as part of this as
well as it would cause the tests to fail othewise.