Moving screen management logic out to a new struct and rework hook triggering
In order to simplify the use and testing of WindowManager, screen
management logic is being moved out to a new struct which will then live
inside of WindowManager. Public API methods from the WindowManager will
delegate through to the methods on the new struct with wrapping around
handling any additional event generated by running the logic in the
screen methods.

This new handing of events also needs to be expanded to cover triggering
hooks so this has now been centralised.

Similar refactoring is going to happen for Workspace and Client
management as well.