Reworking HookName and EventAction to make them serializable
In order to support serializing events for replay, reference types need
to be removed from HookName and EventAction. This results in a breaking
API change to the new_client hook as it how accepts an ID rather than a
mutable reference to the new Client itself. This was always an odd API
to work with (and not in keeping with the other hooks) and the same
functionality is still possible given that the user can simply fetch a
reference to the Client from the window manager reference passed to the
hook if needed.

Examples and built in Hooks have been updated to show the usage of the
new API.