Updating all dependencies
I've not updated the dependencies in any of the Cargo.toml files for
Penrose in quite a while so this is a run of `cargo upgrade` for all
crates and fixing up any associated breaking API changes. A full test
run all passes locally so this should be good to go.