Restructure and moving to simple_test_case
This initial change is largely around sorting out the module structure
(as it currently stands) and imports of macros. There are a couple of
minor changes around internal APIs as well as a move to use
simple_test_case for parameterised tests rather than the clunky
test_cases macro that had been used previously.

The full test suite is still passing so this feels like a good foothold
to start the rewrite. Next up is sorting out the main WindowManager API
so that it is easier to maintain and test which will be a full breaking
change (as opposed to just breaking imports which this commit already
does).