Fixing issue with UnmapNotify handling
When porting over the behaviour from XMonad I ignored (but left a note about...)
some behaviour where the client event mask was updated to remove structure
notify. Turns out, this is pretty important(!)

Without it, the tracking of pending unmaps ends is incorrect and every time a
client is hidden (e.g. when switching workspaces) it actually gets removed from
the ClientSet entirely. This is...not great. You'd like your programs to still
be there when you switch back from another workspace!

The fix is to add a dedicated mask that is set and restored around calling the
XConn unmap method which should _not_ be called directly anywhere else in the
codebase.