Moving state extensions over to being an AnyMap.
Having state extensions as a generic works fine if you only need a single extension,
but composition of multiple state extensions quickly gets messy as each hook that
needs to make use of an extension needs to know how to get at its own data.
Taking a page out of the webserver frameworks, I'm going for a typemap instead so
that hooks can store and fetch their state directly. Building up the extensions is
currently only possible at the WindowManager level due to the extensions being stored
in State rather that Config but that might change in future if that becomes problematic.