NamedScratchpad now works
The main behaviour of named scratchpads is now operational but there are several
bugs outstanding around what the focused workspace / client is which are a little
tricky to nail down at present. I think I'm going to port over the status bar
next as that lets me sort out custom information and tracking focus which is REALLY
helpful for debugging this sort of thing.