Moving built-in but non-core functionality to builtin module
There is a lot of functionality that is provided from the main crate
that isn't strictly the "core" of penrose. It's default implementations
and useful pieces that are worthwhile providing "out of the box". That
kind of functionality it now lives in the builtin module, with the idea
being that "core" really is the core of the whole system.