Fixing issue with manage hooks leading to multiple refreshes
The built-in manage hooks were written to use modify_and_refresh which was
leadin to multiple refreshes being triggered as when we run the manage hook
we then immediately trigger a refresh internally anyway.

This commit also adds some minimal comments in the docs for ManageHook to
point this out for future but it should be fleshed out when everything gets
full docs in the future.