Fixing stacking order bug
When floating windows were present it was clear that the previous impl of restack
wasn't behaving correctly. The tests around the pure logic all seemed sane though
and the behaviour I had in the previous design now results in wild flickering as
the windows fight one another for top position it seems?
Either way, this is now working as intended when a floating window is present on
the screen (both with the window being mapped / unmapped when new clients are
added to the workspace).