Fixing issue with invisible workspaces when redetecting screens
When we redetect the connected screens we backfill focussed workspaces from
the 'hidden' list we maintain in the StackSet. This works fine if there are
no invisible workspaces but given that we were blindly popping the head of
the hidden list we could end up placing an invisible workspace on a new screen
if it was the most recently hidden workspace.

This was most evident when using named scratchpads and connecting a new
monitor while penrose was running. There was also the added issue that the
newly focused workspace was essentially random rather so this change also makes
sure to pull from the available workspaces in ID order.