Fixing broken 'drag_workspace' behaviour
Dragging a focused workspace from one screen to another requires us to be careful
about the previously focused tag (arguably, using a Stack for the screens isn't
_quite_ good enough as we have this idea of previous focus that we are now book
keeping manually...). For the simple test case data I had before, everything was
working fine for both the changes to pure state and what we generated when
diffing around a drag operation. The issue came in when (as will be the case in
actual usage) the previously focused tag is _not_ the tag that is on the other
screen. In that case we ended up mangling what the value of 'previous_tag' was
but in a consistent way that meant you could "correctly" drag to the other screen
if you did it twice with no other operation in between...

This commit extends the existing tests (which the previous behaviour failed on)
and correctly handles the value of 'previous_tag' when dragging.