Remove note about 3.0 and git install in book
The 3.0 note and git flag for the install seems outdated if I'm not mistaken. I'm just getting started and while reading through and went ahead before I saw what the book said. But it seems to give the same version and given that the version seems to be stable now I assume it is fine? Though should mention I haven't gotten it to work with either so far, but since it is neither of them I assume it is not based on this. Thought I might as well throw this up here when I noticed it.