Add support for dmenu-rs and more dmenu options.
Custom font is now supported with -fn, display dmenu on bottom with -b.
Added optional custom prompts to src/extensions/actions/dynamic_select.rs so the user isn't
stuck with a default prompt. Added a new function which just spawns
dmenu with a custom prompt and config, as it seems like a sane
default/nicety to have in the extensions.