Adding LayoutHooks as a feature in Config
So far the way to modify Layouts has been to wrap them in a LayoutTransformer.
This gives you a lot of control over how the underlying Layout is run and lets
you dynamically alter that behaviour as the window manager runs as the
LayoutTransformer trait also supports handling Messages.

However, for the common use case of things like adding gaps between windows or
reserving space at the top or bottom of the screen you end up needing to apply
a copy of the LayoutTransformer to every Layout in every Workspace. A LayoutHook
is a way of running the pre/post transformation logic allowed by a
LayoutTransformer centrally each time _any_ Layout is called, at the expense of
not being able to interact with messages or modify/swap out the underlying
Layout being run.

Implementing this has involved a refactor in how the internal logic around
positioning clients and generating the state diff works as we need to be able to
inject and run the user's LayoutHook (previously this was all pure but now we
are interacting with State and the XConn). The existing tests around the
positioning code have been maintained by adding some cfg(test) helper methods
that plumb through a stub State and XConn that are unused when there is no
LayoutHook to run.

For reference, I've also implemented a built-in SpacingHook which combines the
logic of the Gaps and ReserveTop LayoutTransformers.

TODO: Update the book to cover LayoutHooks.