Fixes #268: explicitly clearing NSP state on client destroyed
So far it has only been observed with the st terminal, but it looks like it is possible
for a newly spawned window to re-use the Xid of a previous client of the same program.
If a NamedScratchpad still thinks it owns a window with that ID it will run the toggle
logic under the assumption that the window in question is the one it owned previously.

The initial fix for this is to add an event hook that checks for NamedScratchpad clients
being destroyed and explicitly removes the stored ID from state. Alternatively we could
also look at running the NSP Query every time the toggle binding is run but this feels
like a better approach as it keeps the toggle logic as simple as possible.