This is initial work on a BSP layout
Aiming for something like i3 in terms of functionality and making use
of a BSP zipper in order to manipulate the tree. At this stage I've
got a working tree zipper and node structs that support splitting,
rotating and traversal, with the first pass at the layout algorithm
itself just being a zip of the client stack with the Rect positions
generated by the tree.

There is still work to do around keeping the state of the tree in
sync with the workspace (currently the tree is completely unaware of
the workspace it is laying out) and I'll need to have a think about
how this interacts with other layouts when switching between them.
Initial thoughts are that I'll probably need to broadcast some messages
to ALL layouts when clients are added and removed from workspaces
as part of modify_and_refresh but we'll see when we get to that point.