Addressing memory leak in status bar
I'm still not 100% confident that this is the last of the memory leak
issues in the status bar but at least locally for me this gets things
under control. There was a build up of un-acked expose / noexpose events
for the status bar being mapped that look to have been the remaining
source of leaked memory. Strangely, In working on this I repeatedly
ended up in a situation where x11rb was building up a large internal vec
(multiple MB) of what I _think_ was cookies for outstanding requests
that it had not received a reply for? I couldn't find any way of
inspecting that vec or forcing it to be cleared so if there continues to
be memory leak issues in this area I will likely need to reach out to
Psychon and check what it is I am missing here.